+++ src/utils.cpp
@@ -8,11 +8,11 @@ int Utils::randInt(int absMax) {
     return rand()%absMax - rand()%absMax;
 }
 
-double Utils::randFloat(int absMax) {
+long double Utils::randFloat(int absMax) {
     int a = randInt(absMax);
     int b;
     do {
         b = randInt(absMax);
     } while (b == 0);
-    return (double)a/b;
+    return (long double)a/b;
 }
